[sfmData] View metadata: add readRealNumber to support fraction
